Improved login + added support for logout.
Caveat: due to a logoff bug in zotonic, if you hit login
right after a logout from a perona session, it'll log you
right back in again (kind of as a 'rememberme' cookie) due
to the fact that the persona session never got a chance to
properly logout first.
The persona session will be logged out at the next page after
a logout, though. Only it takes a few seconds...